moved _is_avail func to gpu.py
